refactor: 폴더 구조 정리 및 개선 (api, stores 폴더)
