+++ docs/codex/store/02-implementation/매장_판매영역_구현_내역.md
@@ -71,7 +71,7 @@
 
 ### 3.2 판매 모델
 
-`src/stores/sales.js`
+`src/stores/storeSales.js`
 
 새 판매 스토어를 추가했다.
 
@@ -241,7 +241,7 @@
 - `src/config/roleMenus.js`
 - `src/router/index.js`
 - `src/stores/inventory.js`
-- `src/stores/sales.js`
+- `src/stores/storeSales.js`
 - `src/views/store/StorePosView.vue`
 - `src/views/store/StoreSalesHistoryView.vue`
 - `src/views/store/StoreSalesAnalysisView.vue`
