+++ src/stores/store/storeAiReport.js
