+++ src/stores/store/storeOrders.js
@@ -1,4 +1,4 @@
-import { useStoreOrderStore } from '@/stores/storeOrder.js'
+import { useStoreOrderStore } from '@/stores/store/storeOrder.js'
 
 // Backward compatibility shim. Prefer useStoreOrderStore directly.
 export const useStoreOrdersStore = () => useStoreOrderStore()
