+++ src/views/hq/circular-stock/HqCircularStockSalesHistoryView.vue
@@ -4,8 +4,8 @@ import { useRouter } from 'vue-router'
 import AppLayout from '@/components/common/AppLayout.vue'
 import { roleMenus } from '@/config/roleMenus.js'
 import { useAuthStore } from '@/stores/auth.js'
-import { useCircularStockStore } from '@/stores/circularStock.js'
-import { useCircularStockBuyerStore } from '@/stores/circularStockBuyers.js'
+import { useCircularStockStore } from '@/stores/hq/circularStock/circularStock.js'
+import { useCircularStockBuyerStore } from '@/stores/hq/circularStock/circularStockBuyers.js'
 
 const router = useRouter()
 const auth = useAuthStore()
