+++ src/views/store/inbound/StoreInboundAnalysisView.vue
@@ -4,7 +4,7 @@ import { useRouter } from 'vue-router'
 import AppLayout from '@/components/common/AppLayout.vue'
 import { roleMenus } from '@/config/roleMenus.js'
 import { useAuthStore } from '@/stores/auth.js'
-import { useStoreInboundStore } from '@/stores/storeInbound.js'
+import { useStoreInboundStore } from '@/stores/store/storeInbound.js'
 
 const router = useRouter()
 const auth = useAuthStore()
